basic ui shell for actors
Added the basic Panel class to be used for all future swing ui work. I
also created a new launch config, in the Window class. This will be
where the actor tests are run from, as it creates all of the child
panels it needs. I added a .gitignore file so that the bin folder is not
constantly being pushed and pulled.