UI to edit actor properties. Added ActorBin
changed ValuesPanel so that it is not constrained to only work for
Action parameters. It can now hold the values of any collection of
objects. The ActorBin is a wrapper to create Actors. The ActorBin keeps
a running log of Actors such that other functions can access any Actor
they please via ActorBin.getInstance().lookUpActor(String name). Also
added two functions to Actor. They get/set the actor properties, but
they don't do it such a type checked way. They are only meant to be used
when you cannot get the compiler to recognize what your property types
are, but you know logically that they are valid.