parser work, shyness
more improvement on the Values Panel. Now there is an outside parser
class that you can register parsers to with incredible ease. I also
added a Location Property, and a shyness trait. If you ask a gangster to
sing and there are other people on his street, he will be less inclined
to agree.